0027792: TKMath, BVH -- Add tesselator API and implement base marching cube (MC) tessellator
The goal is to extend BVH package with complete set of algorithms allowing generation of
distant fields from meshes (or another objects) and tessellation of distant fields. This
allows to implement some geometrical algorithm in quite elegant way. E.g., the offset mesh
for the given mesh can be calculated in the following way: the input mesh is converted into
distant field that represents the mesh in form of implicit function D(x, y, z) = 0. After that,
for generation of offset mesh we need to perform tessellation of implicit function D(x, y, z) - T = 0,
where T is desired offset distance.