0028211: Modeling Algorithms - Boolean fuse operation produces incorrect result
The reason of the regression is a mistake while classification of some face (created by Boolean algorithm and looks like thin ring) as hole. In the fix, another algorithm for classification is used.

Additional corrections:

1. Classes IntTools_FClass2d and BRepTopAdaptor_FClass2d have been joined to the single class BRepTopAdaptor_FClass2d (IntTools_FClass2d has been eliminated but now, BRepTopAdaptor_FClass2d is configured as current version of IntTools_FClass2d).

2. Interface of following classes have been changed: BRepClass_FClass2dOfFClassifier, BRepClass_FClassifier, BRepClass_FaceClassifier, BRepClass_FaceExplorer, BRepClass_FacePassiveClassifier, BRepClass_Intersector, BRepTopAdaptor_FClass2d, CSLib_Class2d, Geom2dHatch_Classifier, Geom2dHatch_Elements, Geom2dHatch_FClass2dOfClassifier, Geom2dHatch_Intersector, TopClass_FaceClassifier, CSLib_Class2d.

3. New DRAW-command "ishole" has been created in BOPTest_LowCommands.cxx file. This command checks if the material of the face is inside (not hole) or outside (hole) of wire enclosed region.

4. Changes in BOPAlgo_PaveFiller::PerformEE() and BOPAlgo_PaveFiller::PerformEF() methods (see BOPAlgo_PaveFiller_3.cxx and BOPAlgo_PaveFiller_5.cxx files): The algorithm of the vertex tolerance computation has been corrected in order to make new vertex cover regions with irregularities (such as Self-intersection).

5. Now, BRepClass_FaceExplorer returns set of different segments (methods Segment(...) and OtherSegment(...) of classes BRepClass_FaceExplorer and Geom2dHatch_Elements have been replaced with ListOfRays(...) method of correspond class) which can be used while classifying. Every segment has some weight, which shows to what extent we can trust this segment in classification process. As result, the classificator returns data obtained from the most reliable segment. These weight are stored in BRepClass_FaceExplorer::RayInfo::myWeight and Geom2dHatch_Elements::RayInfo::myWeight variables.

6. New classes TopClass_RayInfo (describes classification ray joined the classification point with some point in the edge of classified face), TopClass_GeomEdge and TopClass_GeomVertex (store geometrical structure of the edges including pseudo-edges and their vertices) have been created.

7. Method BRepTopAdaptor_FClass2d::PerformInfinitePoint() has been corrected in order to use another (than on the MASTER) point for classification.

8. Method IntTools_EdgeEdge::ComputeLineLine(...) returns intersection point even if the edges have shared vertex. Earlier, it did not it.

9. Creation of pseudo-edges in order to cover gaps in geometrically not-closed (in 2D-space) faces. These face are considered as valid if this gap is covered by some vertex. However, classifier cannot work with "covering vertices". Therefore, new feature (pseudo-edge) has been created. In fact, pseudo-edge is line segment in 2D-space joined two not-closed points.

10. BRepClass_Edge class has been deleted.

11. Check if the section curve cannot be converted to TopoDS_Edge has been added in BOPAlgo_PaveFiller::MakeBlocks() method.

12. Algorithms using face-classifier have been adjusted in order to make them send 3D-tolerance instead of 2D one. I.e. if 3D-representation of classified 2D-point lies ON the edge of the face (the distance from it to the edge is less than the edge tolerance) then it is possible to obtain IN/OUT (not ON) status as result of classification in case when the face tolerance is significantly lesser than the edge tolerance.

13. Since now, following classes apply 3D-tolerance intead of 2D (were applied earlier): Adaptor3d_TopolTool::Classify(...), Adaptor3d_TopolTool::IsThePointOn(...), BRepClass_FClass2dOfFClassifier::Reset(...), BRepClass_FClassifier::BRepClass_FClassifier(...), BRepClass_FClassifier::Perform(...), BRepClass_FaceClassifier::BRepClass_FaceClassifier(...), BRepClass_FaceClassifier::Perform(...), BRepClass_FacePassiveClassifier::Reset(...), BRepClass_Intersector::Perform(...), BRepTopAdaptor_FClass2d::BRepTopAdaptor_FClass2d(...), BRepTopAdaptor_FClass2d::TestOnRestriction(...), BRepTopAdaptor_TopolTool::Classify(...), BRepTopAdaptor_TopolTool::IsThePointOn(...).

14. DRAW-commands returning the resolutions of curve/surface have been created (see "curvresolution" and "surfresolution"). Non-regression test cases have been created for these commands in geometry/props grid.

15. Creation of test cases for issues #28211 and #28800.

16. Code refactoring and documenting.

-----------------------------------------------------------------

Correction of test cases according to their new behavior.

1. tests\bugs\modalg_6\bug27341*

They are not regressions/improvements. Simply another behavior. Bug #27341 is not reproduced.

2. de\iges_2\J1
   de\iges_3\B2.

Source data are not valid for classification algorithm (self-intersected face like attachments faceDeI2J1.brep and FaceDeI3B2.brep).

3. de\step_3\A3
   de\step_3\C3

Invalid (in terms of BRepCheck algorithm) face (let call it "ff") contains edge ff_4 with bending p-curve. Such faces are too hard for classification.

4. tests\de\step_3\D9

The face has invalid p-curve (see FaceDeS3D9.brep attached file).

5. tests\bugs\modalg_6\bug26953_3

Calls classifier for not-closed face. Before the fix, classification point was IN the face. Now, it is out (by the reason of entering of pseudo-edges).

6. tests\bugs\modalg_5\bug23849_2

The reason of the regression is described in item 12.

7. tests\de\step_1\I1

Contains the face having non-manifold structure (see attached picture "de step_1 I1.PNG").

8. bugs iges bug131_4

Difference in the picture is expected. The source face contains really not-oriented wire (see attached picture "bugs iges bug131_4.PNG"). So, the source data is wrong for use in classification algorithm.

9. boolean volumemaker A8

This test case is really unstable. So, we shall apply new behavior. E.g. we cannot classify correctly any point relatively to the face containing only two edges, which are two (almost) coincident lines.

10. blend tolblend_simple C1

The reason while this test has become failure is described in the test case tests\bugs\modalg_6\bug28211_19. Earlier, the classifier returned IN-status. Now, it returns ON. And ON-status is correct result.

11. bugs modalg_6 bug25908

The reason of changes is explained in tests\bugs\modalg_6\bug28211_20. Earlier, the classifier returned OUT-status. Now, it returns ON. And ON-status is correct result.

12. bugs\modalg_2\bug472_1

New result is IMPROVEMENT.

13. tests\prj\base\*

New results are IMPROVEMENT.

14. heal fix_shape A1
    heal fix_shape F7

Before the fix, checkshape reported "Bad orientation" about the result shape. However, indeed, the shape is oriented correctly. So the new behavior of the tests can be considered as IMPROVEMENT. On the other hand, the result shape must not be considered as valid (see issue #0023127).

15. bugs moddata_2 bug377

New result is IMPROVEMENT. Additionally, QA-command "OCC377" has been deleted and the test has been rewritten.

16. de step_2 S1
    de step_3 D9

Since now, some faces are read with another (in compare with the MASTER) orientation (i.e. as a hole). Therefore, we obtain such isolines.