0029093: BOP PaveFiller hungs and constantly consumes memory
Method IntWalk_PWalking::ExtendLineInCommonZone(...) because it seems to be created for single case and (therefore) is not universal. It is bad idea to try to go along iso-line in case of tangent. Now, there is more intellectual method IntWalk_PWalking::PutToBoundary(...) but its current representation allows extending WLine to the boundary only. I.e. if tangent zone is in middle place of the surface this method will have no effect.