0029257: GeomPlate generates surface in unexpected place
Treatment of the parameter EnlargeCoeff in the constructor of GeomPlate_MakeApprox has been corrected. Now the bounds of the result surface are correctly enlarged.

The new draw command 'pullupface' has been implemented to check behavior of GeomPlate.

The test case "perf modalg bug453_2" has been corrected. Now the area is computed the same on both Linux and Windows.