Provide possibility to get the modified faces from the BRepOffsetAPI_PatchFaces algorithm. Two new methods have been added: - to get the new replaced face - TopoDS_Face GetPatchFace(const TopoDS_Face& theFace) const - to get the neighboring faces of the replacing face - void GetAdjacentFaces(const TopoDS_Face& theFace, TopTools_ListOfShape& theNeighbors) const
Both methods take as an IN parameter the original face which was intended to be replaced.
First method returns the new face, on which the original one was replaced, updated to fit the resulting model.
Second method returns the faces adjacent to the original one also updated to fit the resulting model.

The new draw commands have been implemented to test these new methods:
- pfgetpatchface
- pfgetadjacentfaces