Implementation of the Oriented Bounding Boxes using PCA-based algorithm to use in Boolean operations for filtering of the intersections. OBB is implemented in IntTools_OBB class, and is represented as center, axes and half-size dimensions.
The method IntTools_OBB::IsOut(theOther) is based on the Separated Axes theorem for Oriented Bounding Boxes.
According to this theorem it is necessary to consider 15 separating axes.

By default the OBB are not used. To enable the OBB usage in Boolean Operations it is necessary to call the method BOPAlgo_PaveFiller::SetUseOBB(const Standard_Boolean theFlag) with TRUE value.
To enable OBB usage in DRAW it is necessary to use the command:
buseobb 1